requirements.txt
discord.py
discord
guilded.py
guilded
stoat.py
python-dotenv
requests
googletrans==3.1.0a0